% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_counts.R
\name{alignment_counts}
\alias{alignment_counts}
\title{Get alignment counts from BAM files}
\usage{
alignment_counts(inFiles, samtoolsPath = "~/anaconda2/bin/",
  what = "passedQC", verbose = FALSE)
}
\arguments{
\item{inFiles}{Character - BAM file list}

\item{samtoolsPath}{String - path to samtools directory}

\item{what}{String - What type of reads to include}

\item{verbose}{Logical - whether to print progress / results}
}
\value{
Numeric vector? - Read count for each BAM
}
\description{
Given a list of BAM files, return an array of unique (primary) alignments.
}
\details{
Might make sense to give it an option to include, say, multi-mapped alignments.
Requires samtools to be installed.
}
\examples{
readCounts = alignment_counts(bams, verbose=TRUE)
pdf("countsBar.pdf")
barplot(readCounts/1000000, las=2, main="Aligned reads", cex.names = 0.5, names.arg=gsub("_mapped_Aligned.sortedByCoord.out.bam", "", bams), ylab="Millions")
dev.off()
}
\author{
Emma Myers
}
