% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{anno_grch38}
\alias{anno_grch38}
\title{anno_grch38 数据框}
\format{
一个包含 66533 行和 11 列的数据框，具体变量如下：
\describe{
\item{id}{基因的唯一标识符 (字符型)}
\item{eff_length}{基因的有效长度 (数值型)}
\item{gc}{基因的GC含量 (数值型)}
\item{entrez}{基因的Entrez ID (数值型)}
\item{symbol}{基因的符号名称 (字符型)}
\item{chr}{基因所在的染色体 (字符型)}
\item{start}{基因的起始位置 (数值型)}
\item{end}{基因的结束位置 (数值型)}
\item{strand}{基因所在的链 (数值型，1表示正链，-1表示负链)}
\item{biotype}{基因的生物类型 (字符型)}
\item{description}{基因的描述 (字符型)}
}
}
\source{
by function \code{getGeneLengthAndGCContent} from EDASeq package with default parameters at 2023-02-10
}
\usage{
anno_grch38
}
\description{
这个数据框包含了 GRCh38 基因组注释信息，共有 66533 行和 11 列。
}
\keyword{datasets}
