% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnnotation.R
\name{annotate}
\alias{annotate}
\title{Create annotation dataframe}
\usage{
annotate(geneNames, species = "hsapiens_gene_ensembl")
}
\arguments{
\item{geneNames}{vector to annotate}

\item{species}{dataset to use for the annotation}
}
\value{
annotation dataframe
}
\description{
Create annotation dataframe
}
