% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateCounts.R
\name{annotateCounts}
\alias{annotateCounts}
\title{Title Annotate Counts}
\usage{
annotateCounts(counts, anno)
}
\arguments{
\item{counts}{dataframe of counts to be annotated}

\item{anno}{annotation dataframe}
}
\value{
annotated counts
}
\description{
Title Annotate Counts
}
