% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_strand_split.R
\name{bam_strand_split}
\alias{bam_strand_split}
\title{Split BAM with paired-end data into forward-strand and reverse-strand}
\usage{
bam_strand_split(inFiles, samtoolsPath = "~/anaconda2/bin/", outDest = "./",
  verbose = TRUE)
}
\arguments{
\item{inFiles}{Character - BAM file list}

\item{samtoolsPath}{String - path to samtools directory}

\item{outDest}{String - Save new BAM files here}

\item{verbose}{Logical - whether to print progress / results}
}
\description{
Given a list of BAM files, write two new BAM files for each, one containing forward-strand reads and the other reverse-strand reads.
}
\details{
Requires samtools to be installed.
For explanations of samtools flags (like "-f 0x10" and "-f 0x20" below), see this tool:
http://broadinstitute.github.io/picard/explain-flags.html
Note that "-f" means to include only things with these flags, and "-F" means to exclude anything with these flags.
TIME:
}
\examples{
bamfiles = dir('/Volumes/CodingClub1/RNAseq/smallRNA/20180402_NXT/barcodes_new/Emma/filtered_and_indexed', pattern=".bam", full.names=TRUE)
bamfiles = bamfiles[ -which(regexpr(".bai", bamfiles) > 0) ]
bamfiles = bamfiles[c(4,5,6,1,2,3)]
bam_pair_split(bamfiles, outDest=resPath)
}
\author{
Emma Myers
}
