% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToBed.R
\name{bamToBed}
\alias{bamToBed}
\title{Convert BAM files to bed format}
\usage{
bamToBed(filenames, bedtoolsPath = "", outDest = "./", outSuffix = "")
}
\arguments{
\item{filenames}{Character - List of BAM files}

\item{bedtoolsPath}{String - Path to bedtools directory}

\item{outDest}{String - Directory where bed files should be written}

\item{outSuffix}{String - will be appended to original filename (and ".bam" is replaced by ".bed")}
}
\description{
Take a list of BAM files and write a bed file for each one.
}
\details{
Use bedtools function bamToBed convert BAM files to bed format.  Requires bedtools to be installed.
Mitochondrial, unknown, and "random" reads should have been removed already (see sam_filter.R).
If you give the output file a suffix with "outSuffix", you need to start it with a "_" or "." or
whatever if you don't want it just smooshed onto the end of the input file name.
Time:  For NucSeq data, ranged from ~2.5m for cb_p036_1 to ~14.5m for L4_RORb_1.
Example at the command line:
/opt/bedtools2/bin/bamToBed -i samplename.bam > samplename.bed
}
\examples{
samplenames = read.csv('samplenames.txt', comment.char='#')
filenames = paste(samplenames[,1], '_mapped_Aligned.out.bam', sep='')
bamToBed(filenames, bedtoolsPath = '/opt/bedtools2/bin/')
}
\author{
Emma Myers
}
