% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_pca}
\alias{batch_pca}
\title{Produce PCA in batch}
\usage{
batch_pca(
  pca_infos,
  txi,
  metadata = NULL,
  outdir = NULL,
  r_objects = NULL,
  force = FALSE
)
}
\arguments{
\item{pca_infos}{A csv file or a \code{data.frame} describing the PCA to
produce.}

\item{txi}{The \code{txi} object returned by the \code{import_kallisto}
function.}

\item{outdir}{The directory where to save PCA in pdf format. If \code{NULL},
the plots won't be saved as pdf. The files will be saved as
<outdir>/<id_plot>.pdf. Default: \code{NULL}.}

\item{r_objects}{The directory where to save PCA in rds format. If
\code{NULL}, the plots won't be saved as rds files. The files will be saved
as <r_objects>/<id_plot>.rds. Default: \code{NULL}.}

\item{force}{Should the files be re-created if they already exists? Default:
\code{FALSE}.}

\item{metatada}{The metadata to add to the coordinate table to add color or
shape with the results and the \code{plot_pca} function. If there is no
metadata available, keep the default value of \code{NULL}. Value can either
be a csv file or a \code{data.frame}. Default: \code{NULL}.}
}
\value{
Invisibly returns a \code{list} of all the ggplots.
}
\description{
The goal of this function is to take a csv file or a \code{data.frame}
describing all the PCA to produce and launch them in batches.
}
\details{
The table must contain the following columns:
  * id_plot: unique identifier for this graph
  * group: column from the metadata file to subset the txi. If value is
           NA, all the samples will be used in the PCA.
  * group_val: Value in the group column to use in current PCA
  * use_normalisation: "none", "ruvg" or "combat" (txi must be produced in
             consequence)
  * min_counts: Mean values of counts (TPM, ruvg or combat) to keep gene for
                PCA.
  * id_metadata: Column in metadata table that contains the sample names.
  * size: The point size
  * shape: The column in the metadata file to use to define shapes
  * color: The column in the metadata file to use to define colors
  * title: The title of the PCA
  * legend.position: "left", "right", "top" or "bottom"
  * legend.box: "horizontal" or "vertical"
}
\examples{
pca_infos <- get_demo_pca_infos_file()
txi <- get_demo_txi(large = TRUE)
metadata <- get_demo_metadata_file()
gg_list <- batch_pca(pca_infos, txi, metadata)

}
