% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedgraph_flip.R
\name{bedgraph_flip}
\alias{bedgraph_flip}
\title{Convert bedgraph values between positive and negative}
\usage{
bedgraph_flip(inFiles, outDest = "./", outSuffix = "_flipped")
}
\arguments{
\item{inFiles}{Character - Bedgraph file list}

\item{outDest}{String - Directory where output files should be saved}

\item{outSuffix}{String - will be appended to original filename; default "_flipped"}
}
\description{
Get raw counts of reads mapped to specified genomic features using featureCount tool from the subread package.
}
\details{
Take a list of bedgraph files and convert positive values to negative and vice-versa.  This is a hack for getting
value from two different files into the same deeptools heatmap, but with a different color scheme for each
(using computeMatrix() and plotHeatmap()).  Handy for being able to distinguish forward- and reverse-strand reads.
TIME:  3-5 s per bedgraph, at least for ~15-20MB ones.
}
\examples{
bgs = list.files(pattern = "minus.bedgraph")
bedgraph_flip(bgs)
}
\author{
Emma Myers
}
