% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedgraphToBigWig.R
\name{bedgraphToBigWig}
\alias{bedgraphToBigWig}
\title{Bedgraph files to bigwig format}
\usage{
bedgraphToBigWig(filenames, genomeSizeFile, ucscPath = "/opt/UCSCtools/",
  outDest = "./", outSuffix = "")
}
\arguments{
\item{filenames}{Character - List of bedgraph files}

\item{genomeSizeFile}{String - Filename (with path) to text file containing genome size information}

\item{ucscPath}{String - Path to UCSC tools executables}

\item{outDest}{String - Directory where bigwig files should be written}

\item{outSuffix}{String - will be appended to original filename (and ".bedgraph" is replaced by ".bigwig")}
}
\description{
Take a list of bedgraph files and write a bigwig file for each one.
}
\details{
Use UCSC function bedgraphToBigWig to convert bedgraph files to bigwig format.
Requires UCSC tools to be installed.
Time:  ~40s - 1m per bedgraph, where the bedgraphs are generally ~400-800 MB.
Example at the command line:
system2('/opt/UCSCtools/wigToBigWig', args=c('samplename.gencov.bedgraph', '/path/to/genomeSizeFile.txt', samplename.bigwig))
}
\examples{
bedgraphToBigWig(filenames, genomeSizeFile="/path/to/genomeSizeFile.txt")
}
\author{
Emma Myers
}
