% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMatrix.r
\name{computeMatrix}
\alias{computeMatrix}
\title{Read density per region}
\usage{
computeMatrix(bigwigs, regionsFiles, deeptoolsPath = "", matDest = "./",
  outSuffix = "", textMatrix = FALSE, startLabel = NULL,
  endLabel = NULL, skipZeros = FALSE, missingDataAsZero = FALSE,
  maxThreshold = NULL, minThreshold = NULL, upstream = 0,
  downstream = 0, regionBodyLength = 1000, binSize = 10,
  nProcessors = 1)
}
\arguments{
\item{deeptoolsPath}{String - Path to deeptools directory}

\item{outSuffix}{String - will be appended to original filename (and ".bedgraph" is replaced by ".bigwig")}

\item{filenames}{Character - List of BAM files}

\item{genomeSizeFile}{String - Filename (with path) to text file containing genome size information}

\item{bwDest}{String - Directory where bigwig files should be written}
}
\description{
Write matrix containing read density per region from bigwig files
IF ERROR, CHECK WHAT I WAS DOING LIBRARY(TOOLS) FOR
}
\details{
Take a list of bigwigs and use the deeptools computeMatrix function to get a matrix file for each,
which can then be used by plotHeatmap to create an eps.
IMPROVE: You have to be in the directory with the input files right now.  At least if you're giving it a
matDest and maybe if you're not, I dunno.
Make it so you can do reference-point instead of scale-regions.
Could be two different functions if the options are really different.  a and b have different defaults.
Don't bother making it possible to group bigwigs.  They're still separate plots, just in the same output file.
Have it write the deeptools command to the top of the err file.
Time:
~14m for a 90MB bigwig, to 5MB matrix.
Example at the command line (if you want to play with the parameters while looking at just one file, this might be easiest):
computeMatrix scale-regions -S samplename.bw -R All_mm10_wholeGenes.bed -b 1000 -a 1000 --skipZeros -o samplename_regions.mat.gz -p 2
}
\examples{
computeMatrix("samplename.bigwig", regionsFiles=c("All_mm10_wholeGenes.bed", "Shuffled_RORht_mm10.bed"), outSuffix="genebody")
}
\author{
Emma Myers
}
