% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alignments.R
\name{count_alignments}
\alias{count_alignments}
\title{Get alignment counts from BAM files}
\usage{
count_alignments(inFiles, samtoolsPath = "~/anaconda2/bin/",
  what = "primary", verbose = TRUE)
}
\arguments{
\item{inFiles}{Character - BAM file list}

\item{samtoolsPath}{String - path to samtools directory}

\item{what}{String - What type of reads to include}

\item{verbose}{Logical - whether to print progress / results}
}
\value{
Numeric vector? - Alignment count for each BAM
}
\description{
Given a list of BAM files, return an array of either unique (primary) alignments or all alignments that passed QC.
}
\details{
Might make sense to give it an option to include, say, multi-mapped alignments.
Requires samtools to be installed.
For explanations of samtools flags (like "-f 0" and "-F 0x904" below), see this tool:
http://broadinstitute.github.io/picard/explain-flags.html
Note that "-f" means to include only things with these flags, and "-F" means to exclude anything with these flags.
TIME:  ~1m / BAM.
}
\examples{
readCounts = count_alignments(bamFileNames)
pdf("countsBar.pdf")
barplot(readCounts/1000000, las=2, main="Uniquely mapped reads", cex.names = 0.5, names.arg=gsub("_mapped_Aligned.sortedByCoord.out.bam", "", bamFileNames), ylab="Millions")
dev.off()
}
\author{
Emma Myers
}
