% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{count_features}
\alias{count_features}
\title{Get read counts from featureCounts summary files}
\usage{
count_features(inFiles, verbose = TRUE)
}
\arguments{
\item{inFiles}{Character - List of featureCounts output files}

\item{verbose}{Logical - whether to report progress}
}
\value{
Data frame - Columns correspond to input files and rows to read types
}
\description{
Given a list of featureCounts summary files, return a dataframe of read counts broken down by type (assigned, unassigned due to multimapping, etc)
}
\details{
Gets read counts per sample from the "summary.txt" files output by featureCounts.  This means you're getting counts for the features included
in the annotation file given to featureCounts.  This is how to get specifically intronic read counts, by giving featureCounts a gtf with only introns
and then giving this function the resulting summary files.
TIME:
}
\examples{
readCounts = count_features(summaryFiles)
pdf("countsBar.pdf")
barplot(readCounts/1000000, las=2, main="Total reads", cex.names = 0.5, names.arg=gsub("_mapped_Aligned.sortedByCoord.out.bam", "", bams), ylab="Millions")
dev.off()
}
\author{
Emma Myers
}
