% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads.R
\name{count_reads}
\alias{count_reads}
\title{Get read counts from fastqs}
\usage{
count_reads(inFiles, verbose = TRUE)
}
\arguments{
\item{inFiles}{Character - List of fastqs}

\item{verbose}{Logical - whether to print progress / results}
}
\value{
Numeric vector? - Read count for each input file
}
\description{
Given a list of fastq files, return an array of read counts.
}
\details{
If you use (unzipped!) fastqs, this function will call wc -l through the command line to get the number of reads in the file,
and then divide by 4 to get the read count (fastqs include 4 lines for each read).
TIME:  About 1.5s / GB, if you look at the fastq file sizes.  That's on Smintheus.  Roughly 30s per GB, on yasuimac.
}
\examples{
fastqFiles = list.files("NucSeq/trimmed", pattern=".fastq")
readCounts = count_reads(paste("NucSeq/trimmed/", fastqFiles, sep=""))
pdf("NucSeq_reads_per_sample_trimmed.pdf")
barplot(readCounts/1000000, las=2, main="Nuc-seq reads per sample (trimmed files)", cex.names = 0.5, names.arg=gsub("_trimmed_m20_q20.fastq", "", fastqFiles), ylab="Millions")
dev.off()
}
\author{
Emma Myers
}
