% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_check.R
\name{dir_check}
\alias{dir_check}
\title{Check directory existence}
\usage{
dir_check(dirname)
}
\arguments{
\item{dirname}{String - directory to check}
}
\value{
String - directory with slash at end of name (if it didn't have one already)
}
\description{
Make sure a given directory exists, and if there's no slash at the end of the path, add it.
}
\details{
Causes error if specified directory doesn't exist.  Puts slash at end of name if it's not there already.
}
\author{
Emma Myers
}
