% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4 class for DNA.R
\docType{class}
\name{DNAString-class}
\alias{DNAString-class}
\title{DNASequence Class}
\description{
S4 class for representing DNA sequences.
}
\section{Slots}{

\describe{
\item{\code{DNASequence}}{A character string representing DNA sequence.}
}}

