% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_random_picture.R
\name{draw_random_picture}
\alias{draw_random_picture}
\title{Draw a Random Picture}
\usage{
draw_random_picture(N = 15)
}
\arguments{
\item{N}{Number of points to draw.}
}
\value{
Nothing created.  This function is called for it's side effects.
}
\description{
This function uses ggplot to draw a random picture.
}
\examples{
draw_random_picture(N=17)
}
\author{
John J. Gosink
}
