% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ensembl91_tx_2_all}
\alias{Ensembl91_tx_2_all}
\title{Annotation for Ensembl release 91}
\format{A data.frame with 200243 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}}
\usage{
Ensembl91_tx_2_all
}
\description{
Basics annoations for the Ensembl release 91
}
\keyword{datasets}
