% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERCC92}
\alias{ERCC92}
\title{ERCC92 data description}
\format{
A data.frame with 92 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}
}
\usage{
ERCC92
}
\description{
Downloaded from:
https://www.thermofisher.com/order/catalog/product/4456739#/4456739
}
\keyword{datasets}
