% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{expr_counts}
\alias{expr_counts}
\title{Example Gene Expression Counts Data}
\format{
一个数据框，行名为基因，列名为样本。
}
\source{
来自GSE169304的公开数据集。
}
\usage{
expr_counts
}
\description{
这是一个示例基因表达计数矩阵，包含去重后的数据。
}
\examples{
\dontrun{
# 加载示例数据
expr_counts <- read.csv(system.file("extdata", "数据去重-counts.csv", package = "rnaseq"), row.names = 1)
}
}
\keyword{datasets}
