% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprMean.R
\name{exprMean}
\alias{exprMean}
\title{Mean expression heatmaps}
\usage{
exprMean(exprDataFrame, genes = NULL, samples = NULL, L2 = TRUE,
  ylabSize = 8, figHeightPerGene = 20, figWidth = 200,
  colorsPlot = colorRamp(c("turquoise1", "magenta")), ncolors = 5,
  plotTitle = NULL, minVal = NULL, maxVal = NULL, fileOut = NULL)
}
\arguments{
\item{exprDataFrame}{Data frame - Gene x sample expression values (counts, tpm, whatever)}

\item{genes}{Character - Gene symbols that appear as row names in exprDataFrame}

\item{samples}{Character - Sample names that appear as column names in exprDataFrame}

\item{L2}{Logical - Whether to take log2 of values}

\item{ylabSize}{Numeric - Font size for gene symbols}

\item{figHeightPerGene}{Numeric -}

\item{figWidth}{Numeric -}

\item{colorsPlot}{Color ramp -}

\item{ncolors}{Numeric -}

\item{plotTitle}{String -}

\item{minVal}{Numeric -}

\item{maxVal}{Numeric -}

\item{fileOut}{String - If given, save png to with this filename}
}
\value{
Plotly object
}
\description{
Create single column heatmap of mean expression values (across samples)
Needs plotly
}
\details{
Given a gene-by-sample dataframe with expression values, and (optionally) a list of the genes and
samples you want included, make a one-column heatmap of mean expression using plotly.
Note that, probably because I'm not set up to do the paying thing, some stuff doesn't get
incorporated into the output png.
}
\author{
Emma Myers
}
