% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fancy_print}
\alias{fancy_print}
\title{Fancy Print Messages}
\usage{
fancy_print(messages)
}
\arguments{
\item{messages}{A character vector of messages to print.}
}
\value{
None. Messages are printed to the console.
}
\description{
Prints messages with fancy decorations.
}
\examples{
fancy_print(c("Hello", "World"))
}
