% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts_run.R
\name{featureCounts_run}
\alias{featureCounts_run}
\title{Read counts from BAM or SAM files}
\usage{
featureCounts_run(inFiles, annotFile,
  fcPath = "/opt/subread-1.6.0-MacOSX-x86_64/bin/", outDest = "./",
  outSuffix = "", runThreadN = 1, multimappers = FALSE)
}
\arguments{
\item{inFiles}{Character - BAM or SAM file list}

\item{annotFile}{String - Name (with path) of gtf file with annotations to define features}

\item{fcPath}{String - Path to directory with featureCounts executable}

\item{outDest}{String - Directory where output files should be saved}

\item{outSuffix}{String - will be appended to original filename (and followed by "_fcounts.txt")}

\item{runThreadN}{Numeric - How many cores to use}

\item{multimappers}{Logical - Whether to count multi-mapping reads.  All reported alignments will be counted, using the 'NH' tag in the BAMs/SAMs.}
}
\description{
Get raw counts of reads mapped to specified genomic features using featureCount tool from the subread package.
}
\details{
Take a list of SAM or BAM files, and get counts of reads mapped to genomic features in specified annotations file.
Command issued will be written to a text file, particularly so you can confirm the annotation file used (unless its filename gets changed).
TIME:  ~1.5m per BAM.
Example at the command line (if you want to play with the parameters while looking at just one file, this might be easiest):
featureCounts -a $annotFile -O -o $outputFile $inputFile -T $nCores
}
\examples{
annotFile=/Volumes/CodingClub1/STAR_stuff/annotations/mm10_refGene.gtf
annotFile=/Volumes/CodingClub1/STAR_stuff/annotations/mm10_refSeq_introns_geneids.gtf
}
\author{
Emma Myers
}
