% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_checks.R
\name{file_checks}
\alias{file_checks}
\title{Check file existence / extension}
\usage{
file_checks(filename, shouldExist = TRUE, extension = NA, verbose = FALSE)
}
\arguments{
\item{filename}{String - file to check}

\item{shouldExist}{Logical - whether you expect the file to exist; default TRUE}

\item{extension}{String - extension the file should have; default NA}

\item{verbose}{Logical - whether to print progress / results}
}
\value{
Logical - whether file checks out
}
\description{
Checks whether a file exists and returns a message that depends on whether that file was supposed to exist.
It can also check whether the file has a given extension, and return a message based on that too.
}
\details{
Checks whether specified file exists and returns TRUE if the file exists and was supposed to or
doesn't and wasn't supposed to, and FALSE otherwise.  Will also return FALSE if given an extension and the
file doesn't have it.
}
\author{
Emma Myers
}
