% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{format_counts}
\alias{format_counts}
\title{Prepare formated table counts.}
\usage{
format_counts(txi, anno, level)
}
\arguments{
\item{txi}{The txi object returned by the import_kallisto function.}

\item{anno}{The anno used when creating the txi object.}

\item{level}{The level of the analysis (gene or transcript)}
}
\value{
A data.frame with the anno and the merged counts values.
}
\description{
The table contains annotation and the counts (raw_counts/tpm/fpkm).
}
\examples{
txi <- get_demo_txi()
counts <- format_counts(txi, "Hs.Ensembl79", "gene")

}
