% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{format_de}
\alias{format_de}
\title{Prepare formated DE table.}
\usage{
format_de(dds, txi, contrast)
}
\arguments{
\item{dds}{The DESeqDataSet object returned by deseq2_analysis.}

\item{txi}{The txi object returned by the import_kallisto function.}

\item{contrast}{The contrast for the comparison (see ?DESeq2::results).}
}
\value{
A data.frame with the anno and the merged counts values.
}
\description{
The table contains annotation and the DE results.
}
\examples{
txi <- get_demo_txi()
design <- get_demo_design()
dds <- deseq2_analysis(txi, design, ~ group)
de <- format_de(dds, txi, c("group", "A", "B"))

}
