% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{gene_ontology_sets}
\alias{gene_ontology_sets}
\title{Gene Ontology (GO) 数据}
\format{
一个列表，其中每个元素都是一个字符向量：
\describe{
\item{名称}{GO术语名称}
\item{内容}{与该GO术语相关的基因名称向量}
}
}
\source{
Gene Ontology Consortium
}
\usage{
gene_ontology_sets
}
\description{
包含Gene Ontology的生物学过程(BP)、细胞组分(CC)和分子功能(MF)数据。
}
\keyword{datasets}
