% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_ranges_reduce.R
\name{genomic_ranges_reduce}
\alias{genomic_ranges_reduce}
\title{Reduce genomic ranges to set of non-overlapping ranges}
\usage{
genomic_ranges_reduce(inFile, outFile = NULL)
}
\arguments{
\item{inFile}{String - Name of bedfile containing genomic ranges}

\item{outFile}{String - Name to give output bedfile with non-overlapping genomic ranges}
}
\value{
reducedRanges
}
\description{
Given a bedfile containing genomic ranges, collapse overlapping ranges using the GenomicRanges package.
}
\details{
TIME: 10-20s, mostly to read the input file.
}
\examples{
bed1 = "/Users/nelsonlab/Documents/Results_temporarily_here/NucSeq_results/what_was_this/mm10_refSeq_exons_filtered_sorted.bed"
ranges = genomic_ranges_reduce(bed1, outFile="~/Documents/mm10_refSeq_exons_filtered_sorted_reduced.txt")
}
\author{
Emma Myers
}
