% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_anno_df}
\alias{get_anno_df}
\title{Extract counts with annotation data.frame from txi}
\usage{
get_anno_df(txi, col_name)
}
\arguments{
\item{txi:}{The txi object returned from the `import_kallisto` function}

\item{colname:}{The txi object returned from the `import_kallisto` function}
}
\value{
A `data.frame` object.
}
\description{
Extract counts with annotation data.frame from txi
}
\examples{
abundances <- get_demo_abundance_files()
file_anno <- get_demo_anno_file()
txi <- import_kallisto(abundances, anno = file_anno)
raw_counts <- get_anno_df(txi, "raw_count")
tpm <- get_anno_df(txi, "tpm")
fpkm <- get_anno_df(txi, "fpkm")
ruvg <- get_anno_df(txi, "ruvg")

}
