% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_anno_file}
\alias{get_demo_anno_file}
\title{Get demo annotation file}
\usage{
get_demo_anno_file()
}
\value{
The path to the annotation file
}
\description{
Get demo annotation file
}
\examples{
file_anno <- get_demo_anno_file()

}
