% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_design}
\alias{get_demo_design}
\title{Get demo design}
\usage{
get_demo_design()
}
\value{
A data.frame corresponding to the demo txi object. See
        ?get_demo_txi.
}
\description{
Get demo design
}
\examples{
design <- get_demo_design()

}
