% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_kallisto_dir}
\alias{get_demo_kallisto_dir}
\title{Get demo kallisto quant dir}
\usage{
get_demo_kallisto_dir()
}
\value{
The path to kallisto results
}
\description{
Get demo kallisto quant dir
}
\examples{
dir_quant <- get_demo_kallisto_dir()

}
