% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVg.R
\name{get_human_hsk}
\alias{get_human_hsk}
\title{Get the list of default human housekeeping genes.}
\usage{
get_human_hsk()
}
\value{
A \code{vector} with human housekeeping genes symbols
}
\description{
Data from de Jonge *et al.*, 2007:
Evidence Based Selection of Housekeeping Genes
}
\examples{
human_hsk <- get_human_hsk()

}
