% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_ruvg_anno_df}
\alias{get_ruvg_anno_df}
\title{Extract RUVg counts with annotation data.frame from txi}
\usage{
get_ruvg_anno_df(txi)
}
\arguments{
\item{txi:}{The txi object returned from the `import_kallisto` function and
the `ruvg_normalisation` function}
}
\value{
A `data.frame` object.
}
\description{
Extract RUVg counts with annotation data.frame from txi
}
\examples{
txi <- get_demo_txi()
txi_ruv <- ruvg_normalisation(txi)
ruvg <- get_ruvg_anno_df(txi_ruv)

}
