% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_tpm_anno_df}
\alias{get_tpm_anno_df}
\title{Extract TPM with annotation data.frame from txi}
\usage{
get_tpm_anno_df(txi)
}
\arguments{
\item{txi:}{The txi object returned from the `import_kallisto` function}
}
\value{
A `data.frame` object.
}
\description{
Extract TPM with annotation data.frame from txi
}
\examples{
abundances <- get_demo_abundance_files()
txi <- import_kallisto(abundances, anno = "Hs.Ensembl79")
tpm <- get_tpm_anno_df(txi)

}
