% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{group_table}
\alias{group_table}
\title{Example Group Table Data}
\format{
一个数据框，包含样本的分组信息。
}
\source{
来自GSE169304的公开数据集。
}
\usage{
group_table
}
\description{
这是一个示例分组表，包含样本分组信息。
}
\examples{
\dontrun{
# 加载示例数据
group_table <- read.csv(system.file("extdata", "group_169304.csv", package = "rnaseq"))
}
}
\keyword{datasets}
