% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hs.Ensembl101}
\alias{Hs.Ensembl101}
\title{Annotation for Homo sapiens Ensembl release 101}
\format{
A data.frame with 191887 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}
}
\usage{
Hs.Ensembl101
}
\description{
Basics annoations for the Homo sapiens Ensembl release 101
}
\keyword{datasets}
