% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hs.Ensembl98}
\alias{Hs.Ensembl98}
\title{Annotation for Homo sapiens Ensembl release 98}
\format{
A data.frame with 190069 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}
}
\usage{
Hs.Ensembl98
}
\description{
Basics annoations for the Homo sapiens Ensembl release 98
}
\keyword{datasets}
