% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hs.Gencode35}
\alias{Hs.Gencode35}
\title{Annotation for Homo sapiens Gencode release 35}
\format{
A data.frame with 229420 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}
}
\usage{
Hs.Gencode35
}
\description{
Basics annoations for the Homo sapiens Gencode release 35
}
\keyword{datasets}
