% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_kallisto}
\alias{import_kallisto}
\title{Import quantifications from Kallisto}
\usage{
import_kallisto(
  filenames,
  anno = "Hs.Ensembl91",
  txOut = FALSE,
  ignoreTxVersion = FALSE,
  ercc92 = FALSE
)
}
\arguments{
\item{filenames}{Paths to the abundance files.}

\item{anno}{The version of the annotation to use. Default: "Hs.Ensembl91"
Currently available:
* Hs.Gencode19
* Hs.Gencode27
* Hs.Gencode32
* Hs.Gencode35
* Hs.Ensembl79
* Hs.Ensembl91
* Hs.Ensembl95
* Hs.Ensembl97
* Hs.Ensembl98
* Hs.Ensembl100
* Hs.Ensembl101
* Mm.Ensembl91
* Mm.Ensembl92
* Mm.Ensembl94
* Mm.Ensembl97
* Mm.Ensembl99
* Mm.Ensembl100
* Mm.Ensembl101
* Rn.Ensembl76
* Rn.Ensembl79
* Rn.Ensembl92
* Rn.Ensembl98
* Bt.Ensembl99
* Mmu.Ensembl101
* peaux_colonisees}

\item{txOut}{Return counts and abundance at the transcript level. Default:
FALSE}

\item{ignoreTxVersion}{Ignore version of tx. Default = FALSE}

\item{ercc92}{Include ERCC92 annotation when importing. Default = FALSE}
}
\value{
A txi object.
}
\description{
Import quantifications from Kallisto
}
\examples{
abundances <- get_demo_abundance_files()
txi <- import_kallisto(abundances, anno = "Hs.Ensembl79")

}
