% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{kegg_pathway_sets}
\alias{kegg_pathway_sets}
\title{KEGG通路数据}
\format{
一个列表，其中每个元素都是一个字符向量：
\describe{
\item{名称}{KEGG通路名称}
\item{内容}{与该通路相关的基因名称向量}
}
}
\source{
KEGG数据库
}
\usage{
kegg_pathway_sets
}
\description{
包含KEGG（Kyoto Encyclopedia of Genes and Genomes）通路信息的数据集。
}
\keyword{datasets}
