% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{metabolism_signature_sets}
\alias{metabolism_signature_sets}
\title{代谢相关表型}
\format{
一个列表，其中每个元素都是一个字符向量：
\describe{
\item{名称}{代谢相关表型名称}
\item{内容}{构成该表型的基因名称向量}
}
}
\source{
各种生物医学研究文献
}
\usage{
metabolism_signature_sets
}
\description{
包含与细胞代谢相关的基因表型数据。
}
\keyword{datasets}
