% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{misc_signature_collection}
\alias{misc_signature_collection}
\title{表型集合}
\format{
一个列表，其中每个元素都是一个字符向量：
\describe{
\item{名称}{表型集合名称}
\item{内容}{属于该集合的基因名称向量}
}
}
\source{
各种生物医学研究文献和数据库
}
\usage{
misc_signature_collection
}
\description{
包含各种生物学过程和疾病相关的基因表型集合。
}
\keyword{datasets}
