% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mm.Ensembl101}
\alias{Mm.Ensembl101}
\title{Annotation for Mus musculus Ensembl release 101}
\format{A data.frame with 119414 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}}
\usage{
Mm.Ensembl101
}
\description{
Basics annoations for the Mus musculus Ensembl release 100
}
\keyword{datasets}
