% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mmu.Ensembl103}
\alias{Mmu.Ensembl103}
\title{Annotation for Macaca mulatta Ensembl release 103}
\format{
A data.frame with 49537 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}
}
\usage{
Mmu.Ensembl103
}
\description{
Basics annoations for the Bos taurus Ensembl release 103
}
\keyword{datasets}
