% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{msigdb_hallmark_sets}
\alias{msigdb_hallmark_sets}
\title{Hallmark基因集}
\format{
一个列表，其中每个元素都是一个字符向量：
\describe{
\item{名称}{Hallmark基因集名称}
\item{内容}{属于该基因集的基因名称向量}
}
}
\source{
Molecular Signatures Database (MSigDB)
}
\usage{
msigdb_hallmark_sets
}
\description{
包含MSigDB的Hallmark基因集数据。
}
\keyword{datasets}
