% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metadata.R
\name{parse_metadata_for_LO_report}
\alias{parse_metadata_for_LO_report}
\title{Parse metadata LO}
\usage{
parse_metadata_for_LO_report(metadata)
}
\description{
`r lifecycle::badge('experimental')`
Parse metadata and generate a draft for pca_info, volcano_info, report_info
files.
}
