% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaux_colonisees}
\alias{peaux_colonisees}
\title{A custom database (Homo sapiens + Staphylococcus epidermidis)}
\format{
A data.frame with 249446 rows and 5 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
  \item{transcript_type}{Transcript biotype}
}
}
\usage{
peaux_colonisees
}
\description{
Homo sapiens: Ensembl release 97
Staphylococcus epidermidis: Ensembl Genome release 44
}
\keyword{datasets}
