% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_group.R
\name{plot_compare_group}
\alias{plot_compare_group}
\title{创建带有统计显著性的分组比较图}
\usage{
plot_compare_group(data, group_col, group_order = NULL, fill_colors = NULL)
}
\arguments{
\item{data}{包含表达数据和分组信息的数据框。}

\item{group_col}{指定分组列名的字符串。}

\item{group_order}{指定所需的分组水平顺序的字符向量。
如果为NULL，则使用现有顺序。}

\item{fill_colors}{用于分组的颜色向量。如果为NULL，则自动生成。
颜色数量应与分组水平数量相匹配。}
}
\value{
表示分组比较图的ggplot对象。
}
\description{
该函数接受表达数据和分组信息，将数据转换为长格式，
并创建一个带有统计显著性指标的箱线图。
}
