% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_explore.R
\name{plot_groups_pca}
\alias{plot_groups_pca}
\title{PCA分析并绘制带有分组变量的PCA图}
\usage{
plot_groups_pca(
    data = expr_tpm,
    group_data = group_table,
    samples_in_rows = FALSE,
    scale_data = TRUE,
    log_transform = FALSE,
    id_column = "sample",
    group_column = "group",
    geom_individuals = "point",
    color_palette = "Set1",
    repel_labels = FALSE,
    n_components = 50,
    axes = c(1, 2),
    add_ellipses = TRUE
)
}
\arguments{
\item{data}{数值型矩阵或数据框，样本在行，变量在列。}

\item{group_data}{包含样本元数据的数据框，包括样本ID和分组变量。}

\item{samples_in_rows}{逻辑值。如果为\code{TRUE}，则数据中样本在行，变量在列；如果为\code{FALSE}，数据将被转置。}

\item{scale_data}{逻辑值。如果为\code{TRUE}，在执行PCA之前对数据进行缩放。}

\item{log_transform}{逻辑值。如果为\code{TRUE}，对数据应用\code{log2}转换。}

\item{id_column}{字符串。在\code{group_data}中包含样本ID的列名。}

\item{group_column}{字符串。在\code{group_data}中用于分组样本的列名。}

\item{geom_individuals}{字符串。在PCA图中表示个体的几何形状（例如，"point"）。}

\item{color_palette}{字符串。用于着色组的RColorBrewer调色板的名称。}

\item{repel_labels}{逻辑值。如果为\code{TRUE}，使用\code{ggrepel}避免文本标签重叠。}

\item{n_components}{整数。要计算的主成分数量。}

\item{axes}{整数向量。要绘制的主成分。}

\item{add_ellipses}{逻辑值。如果为\code{TRUE}，在组周围添加置信椭圆。}
}
\value{
ggplot对象，表示PCA图。
}
\description{
对提供的数据执行主成分分析（PCA），并根据指定的分组变量（如治疗、条件等）生成PCA图。
}
\details{
该函数使用\code{FactoMineR}包执行PCA，并使用\code{factoextra}包生成PCA图。样本根据指定的分组变量着色。请确保数据和表型数据中的样本ID匹配。
}
\examples{
\dontrun{
# 假设 'expr_tpm' 是您的数据矩阵，'group_table' 是您的分组数据
plot_groups_pca(
  data = expr_tpm,
  group_data = group_table,
  samples_in_rows = FALSE,
  scale_data = TRUE,
  log_transform = FALSE,
  id_column = "sample",
  group_column = "group",
  geom_individuals = "point",
  color_palette = "Set1",
  repel_labels = FALSE,
  n_components = 50,
  axes = c(1, 2),
  add_ellipses = TRUE
)
}

}
\keyword{PCA}
\keyword{主成分分析}
\keyword{分组}
\keyword{可视化}
