% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_pca}
\alias{plot_pca}
\title{Produce a PCA plot from produce_pca results}
\usage{
plot_pca(df, pca)
}
\arguments{
\item{txi}{The txi object returned by the import_kallisto function.}
}
\value{
Returns the \code{ggplot} object
}
\description{
Produce a PCA plot from produce_pca results
}
\examples{
txi <- get_demo_txi()
res_pca <- produce_pca(txi, graph = FALSE)
p <- plot_pca(res_pca$df, res_pca$pca)

}
