% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{plot_signature_parallel}
\alias{plot_signature_parallel}
\title{并行绘制多个基因集的表型评分箱线图}
\usage{
plot_signature_parallel(
  df_long,
  gene_sets,
  output_dir = "results",
  x = "group",
  y = "Score",
  fill = "group",
  palette = c("#2E86C1", "#E74C3C", "#27AE60"),
  group_levels = NULL,
  ...
)
}
\arguments{
\item{df_long}{一个长格式的数据框，包含表型评分。默认包含以下列："ID"、"group"、"GeneSet"、"Score"。}

\item{gene_sets}{要绘制的基因集名称向量。}

\item{output_dir}{保存绘图的目录（默认是"results"）。}

\item{x}{x轴变量的名称（默认是"group"）。}

\item{y}{y轴变量的名称（默认是"Score"）。}

\item{fill}{填充变量的名称（默认是"group"）。}

\item{palette}{用于绘图的颜色向量。}

\item{...}{传递给 \code{\link{plot_signature}} 的其他参数。}

\item{num_cores}{并行运行的核心数量（默认是 40）。}
}
\value{
返回不可见的NULL。
}
\description{
并行创建并保存多个基因集的表型评分箱线图。
}
\details{
该函数使用 \code{furrr::future_map} 并行创建多个基因集的表型评分箱线图，并将图形保存到指定的输出目录。
}
\examples{
# 示例用法：
\dontrun{
df_long <- tidyr::pivot_longer(score_res$signature_scores, -c(ID, group), names_to = "GeneSet", values_to = "Score")

plot_signature_parallel(
  df_long = df_long,
  gene_sets = score_res$stat_tests$GeneSet,
  group_levels = c("Control", "Disorder")
)
}
}
