% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverables.R
\name{produce_deliverables}
\alias{produce_deliverables}
\title{Produce deliverables}
\usage{
produce_deliverables(
  dir_kallisto,
  anno,
  design,
  contrasts,
  dir_output,
  file_type = "h5",
  digits = 4,
  ignoreTxVersion = FALSE,
  use_ruv = FALSE,
  housekeeping_genes = get_human_hsk()
)
}
\arguments{
\item{dir_kallisto}{Directory with Kallisto quantifications}

\item{anno}{The version of the annotation to use. Default: "Hs.Ensembl91"
Currently available:
* Hs.Gencode19
* Hs.Gencode27
* Hs.Gencode32
* Hs.Gencode35
* Hs.Gencode37
* Hs.Ensembl79
* Hs.Ensembl91
* Hs.Ensembl95
* Hs.Ensembl97
* Hs.Ensembl98
* Hs.Ensembl100
* Hs.Ensembl101
* Mm.Ensembl91
* Mm.Ensembl92
* Mm.Ensembl94
* Mm.Ensembl97
* Mm.Ensembl99
* Mm.Ensembl100
* Mm.Ensembl101
* Rn.Ensembl76
* Rn.Ensembl79
* Rn.Ensembl92
* Rn.Ensembl98
* Bt.Ensembl99
* Mmu.Ensembl101
* Mmu.Ensembl103
* peaux_colonisees}

\item{design}{The experimental design
(1st column: sample; 2nd column: group)}

\item{contrasts}{List of contrasts}

\item{dir_output}{Directory where to write outputs}

\item{file_type}{Abundance file format to use (h5 or tsv).}

\item{digits}{Integer indicating the number of decimal places}

\item{ignoreTxVersion}{Should the version number in the IDs be ignored
during import. Default: \code{FALSE}.}

\item{use_ruv}{Use RUVg normalization. Default: \code{FALSE}.}

\item{housekeeping_genes}{A \code{vector} of gene symbols}
}
\value{
Invisibly returns a list with txi_tx, txi_genes, df_tx (PCA),
        df_genes (PCA), design, contrasts, counts and de values.
}
\description{
counts.csv: counts at the gene level and at the transcript level
de_*.csv: DE analysis at the gene level and at the transcript level for
          specified contrasts
}
\examples{
dir_kallisto <- get_demo_kallisto_dir()
contrasts <- list(comp = c("group", "A", "B"))
design <- get_demo_design()
file_anno <- get_demo_anno_file()
produce_deliverables(dir_kallisto,
                     anno = file_anno,
                     design = design,
                     contrast = contrasts,
                     dir_output = ".",
                     file_type = "tsv",
                     ignoreTxVersion = FALSE)

}
