% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_info_tables.R
\name{produce_info_tables}
\alias{produce_info_tables}
\title{Produce tables required for batch analysis from minimal information}
\usage{
produce_info_tables(
  de_infos,
  design_infos,
  metadata_infos,
  volcano_fc_threshold = c(1.5, 3),
  outdir = NULL,
  header_file = "header.qmd",
  r_objects = "r_objects"
)
}
\arguments{
\item{de_infos}{Path to a csv file describing the DE analyses to perform.}

\item{design_infos}{Path to a csv file describing the groups for the
comparisons.}

\item{metadata_infos}{Path to a csv file of sample metadata.}

\item{volcano_fc_threshold}{Numeric vector of fold-change thresholds to use
in volcano plots. Default: \code{c(1.5, 3)}.}

\item{outdir}{The directory where to save produced tables in csv format. The
files will be saved as <outdir>/<name>.csv. If \code{NULL}, the results
won't be saved as csv. Default: \code{NULL}.}

\item{header_file}{Path to the header file that should be added at the top
of the report.}

\item{r_objects}{A \code{character} string corresponding to the directory
containing the rds files that will be used to create the report.
Default: "r_objects".}
}
\value{
Invisibly returns a named \code{list} of all the tables produced.
}
\description{
Produce tables required for batch analysis from minimal information
}
\examples{
\dontrun{}

}
