% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{produce_pca}
\alias{produce_pca}
\title{Produce a PCA plot}
\usage{
produce_pca(txi)
}
\arguments{
\item{txi}{The txi object returned by the import_kallisto function.}
}
\value{
Produce the PCA and silently returns the data.frame used.
}
\description{
Produce a PCA plot
}
\examples{
txi <- get_demo_txi()
\dontrun{
df <- produce_pca(txi)
}

}
