% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverables.R
\name{produce_txi}
\alias{produce_txi}
\title{Produce txi objects}
\usage{
produce_txi(files, anno, ignoreTxVersion = TRUE)
}
\value{
A list with the txi objects at the gene and at the transcript level
}
\description{
Produce the txi objects at the gene and at the transcript level.
}
