% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fcounts.R
\name{read_fcounts}
\alias{read_fcounts}
\title{Read in raw counts}
\usage{
read_fcounts(inFiles, verbose = TRUE)
}
\arguments{
\item{inFiles}{Character - List of featureCounts output files (not the ones that end in "summary" or "report")}

\item{verbose}{Logical - If true, reports name of each file as it's read}
}
\description{
Read counts from featureCounts output into a dataframe.
}
\details{
Output dataframe has columns corresponding to featureCounts output files, and rows corresponding to features (usually genes).
}
\examples{
example here
}
\author{
Emma Myers
}
