% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReprodPackage2018.R
\docType{package}
\name{ReprodPackage2018}
\alias{ReprodPackage2018}
\alias{ReprodPackage2018-package}
\title{Basic Compilation of RR Scripts}
\description{
This package serves to collect scripts and templates to produce
reports in a more reproducible fashion.
}
