% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rn.Ensembl76}
\alias{Rn.Ensembl76}
\title{Annotation for Rattus norvegicus Ensembl release 76}
\format{A data.frame with 29188 rows and 4 columns:
\describe{
  \item{id}{Ensembl transcript id}
  \item{ensembl_gene}{Ensembl gene id}
  \item{symbol}{Gene symbol}
  \item{entrez_id}{Entrez id}
}}
\usage{
Rn.Ensembl76
}
\description{
Basics annoations for the Rattus norvegicus Ensembl release 76
}
\keyword{datasets}
