% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simple Statistics for determination of RNA.R
\name{rna_from_dna}
\alias{rna_from_dna}
\title{RNA Sequence determination}
\usage{
rna_from_dna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A DNA Sequence Object}
}
\value{
A RNA Sequence Object
}
\description{
Determines a RNA Sequence from a DNA sequence
}
