% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4 class for RNA.R
\docType{class}
\name{RNAStrings-class}
\alias{RNAStrings-class}
\title{RNASequence class}
\description{
S4 class for representing RNA sequences.
}
\section{Slots}{

\describe{
\item{\code{RNASequence}}{A character string representing the RNA sequence.}
}}

