% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVg.R
\name{ruvg_normalization}
\alias{ruvg_normalization}
\title{Normalize data with RUVSeq RUVg algorithm}
\usage{
ruvg_normalization(
  txi,
  housekeeping_genes = get_human_hsk(),
  ignoreTxVersion = FALSE
)
}
\arguments{
\item{txi}{The \code{txi} object returned by the \code{import_kallisto}
function.}

\item{housekeeping_genes}{A \code{vector} of gene symbols}

\item{ignoreTxVersion}{Ignore version of tx. Default = FALSE}
}
\value{
The original txi object with a \code{$ruvg_counts} element added.
}
\description{
Normalize data with RUVSeq RUVg algorithm
}
\examples{
txi <- get_demo_txi()
txi_ruv <- ruvg_normalization(txi)

}
