% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{signature_score}
\alias{signature_score}
\title{计算基因表型得分并进行统计检验}
\usage{
signature_score(
  expression_data= expr_fpkm,
  group_data = group_table,
  sample_column = "sample",
  gene_sets = NULL,
  gene_set_name = "HALLMARK",
  method = "ssgsea",
  num_cores = 20,
  min_gene_count = 5,
  group_comparisons = list(c("Normal", "Disorder")),
  output_dir = NULL,
  padjust_threshold = NULL
)
}
\arguments{
\item{expression_data}{数据框，基因表达数据，行为基因，列为样本（例如经过标准化处理的 TPM、FPKM 等数据）。}

\item{group_data}{数据框，包含样本信息，需包含样本标识列和分组列。}

\item{sample_column}{字符串，表示在 \code{group_data} 中作为样本标识的列名（默认“sample”）。}

\item{gene_sets}{列表，可选提供自定义基因集（默认为 NULL）。}

\item{gene_set_name}{字符串，用于标识要使用的内置基因集的名称前缀（默认为“HALLMARK”）。}

\item{method}{字符串，计算表型得分的方法（默认为“ssgsea”）, 'pca'(矩阵不能包含方差为0的变量), 'ssgsea', 'zscore','integration'。}

\item{num_cores}{数值，使用的核心数（默认为 20）。}

\item{min_gene_count}{数值，最少基因数量（默认为 5）。}

\item{group_comparisons}{列表，指定进行统计检验的分组比较（默认为 list(c("Normal", "Disorder")）。}

\item{output_dir}{字符串，结果保存的目录（默认为 NULL，表示不保存）。}

\item{padjust_threshold}{数值，用于根据调整后的 p 值过滤统计检验结果（默认为 NULL）。}
}
\value{
列表，包含表型得分和统计检验结果。
}
\description{
该函数计算基因表型（signature）得分，并在不同组之间进行统计检验。用户可以选择返回或保存结果。
}
\keyword{基因表型,}
\keyword{数据分析}
\keyword{统计检验,}
