% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_info_tables.R
\name{table_from_file}
\alias{table_from_file}
\title{Read a text-delimited or Excel file into a tibble of character columns}
\usage{
table_from_file(path, delim = NULL, sheet = NULL, range = NULL)
}
\arguments{
\item{path}{A single string giving the path to an existing text or Excel file.}

\item{delim}{Either \code{NULL} or a single string to use as a text
delimiter. Ignored if \code{path} is a path to an Excel file. If \code{NULL},
the delimiter will be guessed from the contents of the file.
Default: \code{NULL}}

\item{sheet}{Either \code{NULL}, a single string or a single integer, giving
the name or number of the sheet to read from an Excel file. Ignored if
\code{path} is a path to a text file, or if the sheet is specified using the
\code{range} argument. If \code{NULL}, the first sheet is read.
Default: \code{NULL}}

\item{range}{Either \code{NULL} or a single string specifying the rectangle
of cells to read in an Excel file. Ignored if \code{path} is a path to a text
file. The rectangle can be specified in any of two formats: "D3:G5" or
"R3C4:R5C7" (both examples specify rows 3 to 5 and columns 4 to 7). The name
of the sheet to read, e.g. "mysheet", can be specified with "mysheet!D3:G5"
or "mysheet!R3C4:R5C7". A sheet name specified in this way overrides the
\code{sheet} argument. If \code{range} is \code{NULL}, the rectangle is
guessed from the contents of the file. Default: \code{NULL}}
}
\value{
A tibble with character columns
}
\description{
This function automatically detects whether a file is in a delimited text
format (e.g. csv, csv2, tsv) or Excel format (xls or xlsx) based on its
contents, and then reads it using format-specific parameters, without parsing
the resulting columns.
}
\examples{
de_infos <- get_demo_de_infos_file()
de <- table_from_file(de_infos, delim = ",")

}
