% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_to_gene_fix.R
\name{transcript_to_gene_fix}
\alias{transcript_to_gene_fix}
\title{Replace transcript ids with gene ids in GTF file}
\usage{
transcript_to_gene_fix(gtf, mapFile, outSuffix = "_geneids")
}
\arguments{
\item{gtf}{String - Name of GTF with transcript ids in gene_id field}

\item{mapFile}{String - Name of csv with gene ids in the first column and transcript ids in the second}
}
\description{
Given a GTF file with transcript ids in the "gene_id" field and a file mapping these to gene ids, write a new GTF with gene ids in that field.
}
\details{
Use the transcript_to_gene_map function to create the mapFile input.
TIME:  Most of it's the "Replacing transcript ids with gene ids in key-value field" part, which is anywhere from 15-30m.
}
\examples{
transcript_to_gene_fix("/Volumes/CodingClub1/RNAseq/Metadata/mm10_refSeq_introns.gtf", "/Volumes/CodingClub1/RNAseq/Metadata/RefSeq_id_map.csv")
}
\author{
Emma Myers
}
