% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_to_gene_map.R
\name{transcript_to_gene_map}
\alias{transcript_to_gene_map}
\title{Create file associating transcript ids with gene ids}
\usage{
transcript_to_gene_map(inFile, outFile)
}
\arguments{
\item{inFile}{String - Name of GTF with transcript ids in gene_id field}

\item{outFile}{String - Name of csv with gene ids in the first column and transcript ids in the second}
}
\value{
gene_transcript
}
\description{
Given a GTF file with both gene ids and transcript ids, write a csv with gene ids in the first column and transcript ids in the second.
}
\details{
Note that gene ids (in the first column of the output file) will be repeated when associated with multiple transcript ids.
TIME: A minute or two.
}
\examples{
ids = transcript_to_gene_map("/Volumes/CodingClub1/RNAseq/Metadata/mm10_refGene.gtf", "/Volumes/CodingClub1/RNAseq/Metadata/RefSeq_id_map.csv")
}
\author{
Emma Myers
}
