% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{tumor_signature_sets}
\alias{tumor_signature_sets}
\title{肿瘤相关表型}
\format{
一个列表，其中每个元素都是一个字符向量：
\describe{
\item{名称}{肿瘤相关表型名称}
\item{内容}{构成该表型的基因名称向量}
}
}
\source{
各种肿瘤研究文献和数据库
}
\usage{
tumor_signature_sets
}
\description{
包含与各种肿瘤类型和肿瘤生物学过程相关的基因表型。
}
\keyword{datasets}
