% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_counts.R
\name{alignment_counts}
\alias{alignment_counts}
\title{Get alignment counts from BAM files}
\usage{
alignment_counts(filenames, samtoolsPath = "", what = "passedQC",
  verbose = FALSE)
}
\arguments{
\item{filenames}{Character - BAM file list}

\item{samtoolsPath}{String - path to samtools directory}

\item{what}{String - What type of reads to include}

\item{verbose}{Logical - whether to print progress / results}
}
\value{
Numeric vector? - Read count for each BAM
}
\description{
Given a list of BAM files, return an array of unique (primary) alignments.
}
\details{
I'm leaving the function name ambiguous because it might make sense to give it an option to include,
say, multi-mapped alignments.
Requires samtools to be installed.
}
\examples{
uniqueReads = alignment_counts(c("S1.bam", "S2.bam"), samtoolsPath="/opt/samtools-0.1.18")
}
\author{
Emma Myers
}
