% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMatrix.R
\name{computeMatrix}
\alias{computeMatrix}
\title{Read density per region}
\usage{
computeMatrix(inFiles, regionsFiles, deeptoolsPath = "", outDest = "./",
  outSuffix = "", textMatrix = FALSE, startLabel = NULL,
  endLabel = NULL, skipZeros = FALSE, missingDataAsZero = FALSE,
  maxThreshold = NULL, minThreshold = NULL, upstream = 0,
  downstream = 0, regionBodyLength = 1000, binSize = 10,
  nProcessors = 1)
}
\arguments{
\item{inFiles}{Character - List of BAM files}

\item{deeptoolsPath}{String - Path to deeptools directory}

\item{outDest}{String - Directory where matrix files should be written}

\item{outSuffix}{String - will be appended to original filename (and ".bedgraph" is replaced by ".bigwig")}

\item{genomeSizeFile}{String - Filename (with path) to text file containing genome size information}
}
\description{
Write matrix containing read density per region from bigwig files
}
\details{
Take a list of bigwigs and use the deeptools computeMatrix function to get a matrix file for each,
which can then be used by plotHeatmap to create an eps.
TIME:  Varies a lot, maybe based on how many cores you're using?  15-35m per bigwig is typical.
COMMAND LINE EXAMPLE (if you want to play with the parameters while looking at just one file, this might be easiest):
computeMatrix scale-regions -S samplename.bw -R All_mm10_wholeGenes.bed -b 1000 -a 1000 --skipZeros -o samplename_regions.mat.gz -p 2
IMPROVE:
Have it write the deeptools command to the top of the file where text printed to screen gets sent.
Make it so you can do reference-point instead of scale-regions.
Could be two different functions if the options are really different.  a and b have different defaults.
Don't bother making it possible to group bigwigs.  They're still separate plots, just in the same output file.
}
\examples{
computeMatrix("samplename.bigwig", regionsFiles=c("All_mm10_wholeGenes.bed", "Shuffled_RORht_mm10.bed"), outSuffix="genebody")
}
\author{
Emma Myers
}
