% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{count_features}
\alias{count_features}
\title{Read in raw counts for genomic features}
\usage{
count_features(inFiles, verbose = TRUE)
}
\arguments{
\item{inFiles}{Character - List of featureCounts output files (not the ones that end in "summary" or "report")}

\item{verbose}{Logical - If true, reports name of each file as it's read}
}
\description{
Read counts from featureCounts output into a dataframe.
}
\details{
Output dataframe has columns corresponding to featureCounts output files, and rows corresponding to features (usually genes).
}
\examples{
fcs = dir(paste(projectPath, "counts/exonic_counts/", sep=""), pattern="_fcounts.txt")
fcs = fcs[ which(regexpr("summary", fcs) < 0) ]
counts_nucseq_ex = count_features(paste(projectPath, "counts/exonic_counts/",fcs,sep=""))
dim(counts_nucseq_ex)
[1] 24746     7
}
\author{
Emma Myers
}
