% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads.R
\name{count_reads}
\alias{count_reads}
\title{Get read counts from fastqs}
\usage{
count_reads(inFiles, what = "all", verbose = FALSE)
}
\arguments{
\item{inFiles}{Character - List of fastqs}

\item{verbose}{Logical - whether to print progress / results}
}
\value{
Numeric vector? - Read count for each input file
}
\description{
Given a list of fastq files, return an array of read counts.
}
\details{
If you use (unzipped!) fastqs, this function will call wc -l through the command line to get the number of reads in the file,
and then divide by 4 to get the read count (fastqs include 4 lines for each read).
TIME:
}
\examples{
readCounts = count_reads(reportFiles)
pdf("countsBar.pdf")
barplot(readCounts/1000000, las=2, main="Total reads", cex.names = 0.5, names.arg=gsub("_mapped_Aligned.sortedByCoord.out.bam", "", bams), ylab="Millions")
dev.off()
}
\author{
Emma Myers
}
