% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_tpm.R
\name{counts_to_tpm}
\alias{counts_to_tpm}
\title{Get transcripts per million}
\usage{
counts_to_tpm(countMat, geneLengths)
}
\arguments{
\item{countMat}{Matrix? -}

\item{geneLengths}{Vector? -}
}
\value{
Matrix - TPM values
}
\description{
Convert read counts to transcripts per million
}
\details{
# Given a gene (or whatever) by sample read count matrix and a length for each gene, return TPM matrix.
 Follows http://www.rna-seqblog.com/rpkm-fpkm-and-tpm-clearly-explained/
 To spell it out a bit more:
 Gene length / 1000 gives you how many kilobases long the gene is.
 Count / length-in-kilobases gives you reads per kilobase for that gene.
  So you've normalized for gene length.
  Now, get how many millions of reads there are in the sample.
  And normalize by that.  Reads per kilobase, per million reads in the sample.
}
\author{
Emma Myers
}
