% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt_run.R
\name{cutadapt_run}
\alias{cutadapt_run}
\title{Trim reads in fastq files}
\usage{
cutadapt_run(readFilesIn, adapters = list(TruSeq_Universal =
  "AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGTAGATCTCGGTGGTCGCCGTATCATT", TruSeq_Index =
  "AGATCGGAAGAGCACACGTCTGAACTCCAGTCACNNNNNNATCTCGTATGCCGTCTTCTGCTTG"),
  cutadaptPath = "~/.local/bin/", outDest = "./", outSuffix = "",
  qualityCutoff = c(0, 0), minLen = 0, nAdapt = 1, trimn = FALSE)
}
\arguments{
\item{readFilesIn}{Character - Fastq file list (can be gzipped)}

\item{adapters}{List - Each element is an adapter sequence; element names are adapter names}

\item{cutadaptPath}{String - Path to directory with cutadapt executable}

\item{outDest}{String - Directory where output files should be saved}

\item{outSuffix}{String - will be appended to original filename (and followed by "_trimmed")}

\item{qualityCutoff}{Numeric - If one value, trim bases with lower quality score than this from 3' end; if two csv values, trim from 3' and 5' ends respectively.  Pre-adpater removal.}

\item{minLen}{Numeric - Reads shorter than this will be tossed}

\item{nAdapt}{Numeric - Cutadapt will assume there could be as many adapters as this on a given read}

\item{trimn}{Logical - Whether to trim flanking Ns (unknown bases)}
}
\description{
Trim reads using cutadapt.  Written using cutadapt v1.16.
}
\details{
Cutadapt's report, normally displayed in the terminal, goes to originalFileName_report.txt.
That report includes the command line parameters, so whatever you use in this script as the adapter sequences, quality cutoff, etc, will appear there.
TIME:
For 5-6 GB fastq files, about 15m.
For 9-10 GB, closer to 30m.
For all the RORb files, total ~7.5hrs.
Example at the command line (if you want to play with the parameters while looking at just one file, this might be easiest):
TESTED THIS
~/.local/bin/cutadapt -a TruSeq_Index=AGATCGGAAGAGCACACGTCTGAACTCCAGTCACNNNNNNATCTCGTATGCCGTCTTCTGCTTG BF_RORbHTp2_1.fastq -o BF_RORbHTp2_1_trimmed.fastq --trim-n -q 20,20 -m 20 -n 3 > BF_RORbHTp2_1_report.txt
}
\examples{
TESTED THIS
fqs = list.files( paste(projectPath,"raw/",sep=""), pattern=".fastq" )
cutadapt_run(paste(projectPath, "raw/", fqs, sep=""), outDest=paste(projectPath, "trimmed/", sep=""), qualityCutoff=c(20,20), minLen=20, nAdapt=3, trimn=TRUE)
}
\author{
Emma Myers
}
