% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{format_counts}
\alias{format_counts}
\title{Prepare formated table counts.}
\usage{
format_counts(txi)
}
\arguments{
\item{txi}{The txi object returned by the import_kallisto function.}
}
\value{
A data.frame with the anno and the merged counts values.
}
\description{
The table contains annotation and the counts (raw_counts/tpm/fpkm).
}
\examples{
txi <- get_demo_txi()
counts <- format_counts(txi)

}
