% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_anno_df}
\alias{get_anno_df}
\title{Extract FPKM with annotation data.frame from txi}
\usage{
get_anno_df(txi, col_name)
}
\arguments{
\item{txi:}{The txi object returned from the `import_kallisto` function}

\item{colname:}{The txi object returned from the `import_kallisto` function}
}
\value{
A `data.frame` object.
}
\description{
Extract FPKM with annotation data.frame from txi
}
\examples{
abundances <- get_demo_abundance_files()
txi <- import_kallisto(abundances, anno = "Hs.Ensembl79")
fpkm <- get_fpkm_anno_df(txi)

}
