% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_raw_count_anno_df}
\alias{get_raw_count_anno_df}
\title{Extract raw counts with annotation data.frame from txi}
\usage{
get_raw_count_anno_df(txi)
}
\arguments{
\item{txi:}{The txi object returned from the `import_kallisto` function}
}
\value{
A `data.frame` object.
}
\description{
Extract raw counts with annotation data.frame from txi
}
\examples{
abundances <- get_demo_abundance_files()
file_anno <- get_demo_anno_file()
txi <- import_kallisto(abundances, anno = file_anno)
raw_counts <- get_raw_count_anno_df(txi)

}
