% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_pca}
\alias{plot_pca}
\title{Produce a PCA plot from produce_pca results}
\usage{
plot_pca(res_pca, color = NULL)
}
\arguments{
\item{res_pca}{The \code{list} returned from the produce_pca function}

\item{color}{The name of the column in \code{res_pca$df} to use to color the
points in the PCA. If \code{NULL}, won't add color. Default: \code{NULL}.}
}
\value{
Returns the \code{ggplot} object
}
\description{
Produce a PCA plot from produce_pca results
}
\examples{
txi <- get_demo_txi()
res_pca <- produce_pca(txi, graph = FALSE)
p <- plot_pca(res_pca)

}
