% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Read density per region - heatmap}
\usage{
plotHeatmap(matrixFiles, deeptoolsPath = "", outDest = "./",
  outSuffix = "", textMatrix = FALSE, interpolationMethod = NULL,
  dpi = NULL, sortRegions = NULL, averageTypeSummaryPlot = NULL,
  missingDataColor = NULL, colorMap = NULL, zMin = NULL, zMax = NULL,
  heatmapHeight = NULL, heatmapWidth = NULL, startLabel = NULL,
  endLabel = NULL, samplesLabels = NULL, regionsLabel = NULL)
}
\arguments{
\item{matrixFiles}{Character - List of mat.gz files generated by computeMatrix()}

\item{deeptoolsPath}{String - Path to deeptools directory}

\item{outDest}{String - Directory where eps files should be saved}

\item{outSuffix}{String - will be appended to original filename}

\item{textMatrix}{Logical - Whether to write matrix to text file}

\item{interpolationMethod}{String - 'nearest' works for our data so far}

\item{dpi}{Numeric -}

\item{sortRegions}{String -}

\item{missingDataColor}{String -}

\item{colorMap}{String -}

\item{zMin}{Numeric -}

\item{zMax}{Numeric -}

\item{heatmapHeight}{Numeric -}

\item{heatmapWidth}{Numeric -}

\item{startLabel}{String -}

\item{endLabel}{String -}

\item{regionsLabel}{Charater -}

\item{averageTypeSummaryProfile}{String -}

\item{samplesLabel}{Character -}
}
\description{
Create heatmap with summary profile from matrix generated by computeMatrix()
}
\details{
Take a list of mat.gz files containing read density for a given set of regions,
created by the deeptools computeMatrix tool (called by computeMatrix() in this package),
and use deeptools plotHeatmap to create an eps with a heatmap and summary profile for each one.
Note that you can't give plotHeatmap a missingDataColor value using #ffffff format when calling
it through R.  So far as I can tell the # is getting interpreted as a comment, and it thinks you're
not giving it anything.
The latest version of deeptools by default uses interpolation method "bilinear".
However, it fails for our data (at least the TTX and the small RNA), whereas "nearest" works.
TIME: ~10s per matrix file
COMMAND LINE EXAMPLE (if you want to play with the parameters while looking at just one file, this might be easiest):
plotHeatmap -m samplename_regionsName.mat.gz -out samplename_regionsName.eps
IMPROVE: If given sample and region labels, some kind of check to see that they correspond to what's
in the matfile somehow.  And check the length of the character vectors to make sure it matches the
length of matrixFiles.
}
\examples{
mats = list.files(pattern='.mat.gz')
plotHeatmap(mats, samplesLabels=sub('.*_', '', sub('_Aligned.*', '', matfiles)))
}
\author{
Emma Myers
}
