% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverables.R
\name{produce_de}
\alias{produce_de}
\title{Produce DE table}
\usage{
produce_de(
  txi,
  design,
  contrasts,
  formula = ~group,
  use_ruv = FALSE,
  ignoreTxVersion = TRUE,
  digits = 4,
  ncores = 1
)
}
\value{
A data.frame with the gene and tx DE merged in a single table
}
\description{
The differential expression (DE) table will be produced by merging results
at the gene and transcripts level.
}
