% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverables.R
\name{produce_deliverables}
\alias{produce_deliverables}
\title{Produce deliverables}
\usage{
produce_deliverables(dir_kallisto, anno, design, contrasts, dir_output,
  file_type = "h5")
}
\arguments{
\item{dir_kallisto}{Directory with Kallisto quantifications}

\item{anno}{The version of the annotation to use. Default: "Hs.Ensembl91"
Currently available:
* Hs.Ensembl91
* Hs.Ensembl79}

\item{design}{The experimental design
(1st column: group; 2nd column: sample)}

\item{contrasts}{List of contrasts}

\item{dir_output}{Directory where to write outputs}

\item{file_type}{Abundance file format to use (h5 or tsv).}
}
\value{
Invisibly returns a list with counts and de.
}
\description{
counts.csv: counts at the gene level and at the transcript level
de_*.csv: DE analysis at the gene level and at the transcript level for
          specified contrasts
}
\examples{
dir_kallisto <- get_demo_kallisto_dir()
contrasts <- list(comp = c("group", "A", "B"))
design <- get_demo_design()
produce_deliverables(dir_kallisto,
                     anno = "Hs.Ensembl79",
                     design = design,
                     contrast = contrasts,
                     dir_output = ".",
                     file_type = "tsv")

}
