% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverables.R
\name{produce_deliverables}
\alias{produce_deliverables}
\title{Produce deliverables}
\usage{
produce_deliverables(dir_kallisto, anno, design, contrasts, dir_output,
  file_type = "h5", digits = 4)
}
\arguments{
\item{dir_kallisto}{Directory with Kallisto quantifications}

\item{anno}{The version of the annotation to use. Default: "Hs.Ensembl91"
Currently available:
* Hs.Gencode27
* Hs.Ensembl91
* Hs.Ensembl79
* Mm.Ensembl91
* Mm.Ensembl92
* Mm.Ensembl94
* Rn.Ensembl76
* Rn.Ensembl92}

\item{design}{The experimental design
(1st column: group; 2nd column: sample)}

\item{contrasts}{List of contrasts}

\item{dir_output}{Directory where to write outputs}

\item{file_type}{Abundance file format to use (h5 or tsv).}

\item{digits}{Integer indicating the number of decimal places}
}
\value{
Invisibly returns a list with txi_tx, txi_genes, df_tx (PCA),
        df_genes (PCA), design, contrasts, counts and de values.
}
\description{
counts.csv: counts at the gene level and at the transcript level
de_*.csv: DE analysis at the gene level and at the transcript level for
          specified contrasts
}
\examples{
dir_kallisto <- get_demo_kallisto_dir()
contrasts <- list(comp = c("group", "A", "B"))
design <- get_demo_design()
produce_deliverables(dir_kallisto,
                     anno = "Hs.Ensembl79",
                     design = design,
                     contrast = contrasts,
                     dir_output = ".",
                     file_type = "tsv")

}
