% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{produce_pca}
\alias{produce_pca}
\title{Produce a PCA plot from txi}
\usage{
produce_pca(txi, graph = TRUE)
}
\arguments{
\item{txi}{The \code{txi} object returned by the \code{import_kallisto}
function.}

\item{graph}{Produce the graph. \code{TRUE} or \code{FALSE}. Default:
\code{TRUE}.}
}
\value{
Produce the PCA and silently returns the \code{data.frame} and the
result from the \code{PCA} function as a \code{list} with 2 elements
(\code{df} and \code{pca}).
}
\description{
Produce a PCA plot from txi
}
\examples{
txi <- get_demo_txi()
df <- produce_pca(txi, graph = FALSE)

}
