% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR_run.R
\name{STAR_run}
\alias{STAR_run}
\title{Map reads to genome}
\usage{
STAR_run(readFilesIn, genomeDir, starPath = "/opt/STAR/bin/MacOSX_x86_64/",
  outDest = "./", outSuffix = "", runThreadN = 1, settings = NULL,
  outSAMtype = "SAM", outFilterType = "Normal",
  outFilterMultimapNmax = 10, alignSJoverhangMin = 5,
  alignSJDBoverhangMin = 3, outFilterMismatchNmax = 10,
  outFilterMismatchNoverReadLmax = 1, alignIntronMin = 21,
  alignIntronMax = 0, alignMatesGapMax = 0, quantMode = NULL)
}
\arguments{
\item{readFilesIn}{Character - name(s) (with path) of the files containing the sequences to be mapped (e.g. RNA-seq FASTQ files)}

\item{genomeDir}{String - Path to directory where genome indices were generated by STAR}

\item{starPath}{String - Path to directory with STAR executable}

\item{outDest}{String - Directory where output files should be saved}

\item{outSuffix}{String - will be appended to original filename}

\item{runThreadN}{Numeric - How many cores to use}

\item{settings}{String - overrides values for all the following parameters by using ENCODE settings for long or short RNA}

\item{outSAMtype}{String - format for output alignment files (SAM, BAM, BAM SortedByCoordinate, BAM Unsorted SortedByCoordinate)}

\item{outFilterType}{String - reduces the number of "spurious" junctions}

\item{outFilterMultimapNmax}{Numeric - max number of multiple alignments allowed for a read: if exceeded, the read is considered unmapped}

\item{alignSJoverhangMin}{Numeric - minimum overhang for unannoted junctions}

\item{alignSJDBoverhangMin}{Numeric - minimum overhang for annotated junctions}

\item{outFilterMismatchNmax}{Numeric - maximum number of mismatches per pair, large number switches off this filter}

\item{outFilterMismatchNoverReadLmax}{Numeric - max number of mismatches per pair relative to read length: for 2x100b, max number of mismatches is 0.04*200=8 for the paired read}

\item{alignIntronMin}{Numeric - minimum intron length}

\item{alignIntronMax}{Numeric - maximum intron length}

\item{alignMatesGapMax}{Numeric - maximum genomic distance between mates}

\item{quantMode}{String - Get either read counts or transcript coordinates ("GeneCounts" or "TranscriptomeSAM")}
}
\description{
Map reads from fastq files to genome using STAR. Most parameter descriptions are from STAR manual.
}
\details{
Take a list of fastq files containing reads, and get alignments with STAR.  Default parameter values are defaults for STAR 2.5.3.
See manual for details on those.  Override by defining "settings" as "ENCODE_long" or "ENCODE_short" to use ENCODE settings for long or short RNA.
THIS WILL BE FASTER IF I COLLAPSE INPUT FILES AND DO JUST ONE STAR COMMAND.
Because it'll only load the reference genome once.
TIME:
Roughly 10m per 5G data file.
5h 30m for all 28 RORb data files.
4h 45m for the 15 EMX-Cre data files.
3hr 30m for the 14 PV data files.
~30m for the Nuc-seq data (12 files?), minus one file.
 Example at the command line using ENCODE settings for long RNA
(if you want to play with the parameters while looking at just one file, this might be easiest):
/opt/STAR/bin/MacOSX_x86_64/STAR --genomeDir $genomeDir --readFilesIn $pathTrimmed$sn$trimmedSuffix --outFilterType BySJout --outFilterMultimapNmax 20 --alignSJoverhangMin 8 --alignSJDBoverhangMin 1 --outFilterMismatchNmax 999 --outFilterMismatchNoverLmax 0.04 --alignIntronMin 20 --alignIntronMax 1000000 --alignMatesGapMax 1000000 --outSAMtype 'BAM SortedByCoordinate' --outFileNamePrefix $pathMapped$sn$mappedSuffix#'
}
\examples{
genomeDir='/Volumes/CodingClub1/STAR_stuff/indexes/generated_with_gtf_from_mysql/' # generated to be ideal for files w/ max read len 75 (or 76?)
}
\author{
Emma Myers
}
